refactor(solver): extract stats helpers to observability.py
